Moved the test controller into Test direcotoy in controllers
